Add state and state updating functions for enhancer region track updates (#19)
Only does so for arc track for now for testing purposes -- stacked bar and stratified bed will come after

* Dedupe default variant track definition between view config and variant track state

This sets a pattern for doing so with enhancer genes as well.

* Add new query syncer for enhancer gene track

* Add function to update enhancer region track, for arcs (not stacked bar yet)

* clean up and format
