+++ src/components/enhancer-regions/EnhancerRegions.js
@@ -41,6 +41,7 @@ import {
   useEnhancerRegionsShowInfos,
 } from '../../state/enhancer-region-state';
 import { variantTracksState } from '../../state/variant-track-state';
+import { enhancerRegionsTrackState } from '../../state/enhancer-region-state';
 
 import {
   updateViewConfigFocusGene,
@@ -57,7 +58,9 @@ import {
   IGNORED_FOCUS_ELEMENTS,
 } from '../../constants';
 import {
-  DEFAULT_VIEW_CONFIG_ENHANCER,} from '../../view-config-types'
+  DEFAULT_VIEW_CONFIG_ENHANCER,
+  updateViewConfigEnhancerRegionTracks,
+} from '../../view-config-typed';
 
 import { chrRangePosEncoder } from '../../utils';
 
@@ -179,6 +182,7 @@ const EnhancerRegion = React.memo((props) => {
   const setHiglass = useSetRecoilState(higlassEnhancerRegionsState);
 
   const variantTracks = useRecoilValue(variantTracksState);
+  const enhancerTrackConfig = useRecoilValue(enhancerRegionsTrackState);
   const hideUnfocused = useRecoilValue(enhancerRegionsHideUnfocusedState);
   const variantYScale = useRecoilValue(variantYScaleState);
   const colorEncoding = useRecoilValue(enhancerRegionsColorEncodingState);
@@ -249,7 +253,8 @@ const EnhancerRegion = React.memo((props) => {
           force: shouldSkipUpdatingXDomain,
         }),
         updateViewConfigFilter(selectedSamples),
-        updateViewConfigMatrixHeight(numSamples)
+        updateViewConfigMatrixHeight(numSamples),
+        updateViewConfigEnhancerRegionTracks(enhancerTrackConfig)
       )(deepClone(DEFAULT_VIEW_CONFIG_ENHANCER)),
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [
