+++ src/state/variant-track-state.ts
@@ -4,7 +4,13 @@ import {
   useRecoilQueryString,
   useRecoilQueryStringSyncher,
 } from '../utils/query-string';
-import { getDefault } from './utils';
+import { Track, TrackType } from '../view-config-types';
+import {
+  getDefault,
+  TrackSourceAbbr,
+  TRACK_SOURCE_ABBR_TO_SERVER_URL,
+  SERVER_URL_TO_TRACK_SOURCE_ABBR,
+} from './utils';
 
 type VariantTrack = {
   server: string;
@@ -15,31 +21,16 @@ type VariantTrack = {
   label: string;
 };
 
-const enum VariantTrackAbbr {
-  RG = 'rg',
-  HG = 'hg',
-}
-
-const VARIANT_TRACK_ABBR_TO_SERVER: Record<VariantTrackAbbr, string> = {
-  [VariantTrackAbbr.RG]: 'https://resgen.io/api/v1',
-  [VariantTrackAbbr.HG]: 'https://higlass.io/api/v1',
-};
-
-const VARIANT_TRACK_SERVER_TO_ABBR: Record<
-  string,
-  VariantTrackAbbr
-> = Object.fromEntries(
-  Object.entries(VARIANT_TRACK_ABBR_TO_SERVER).map((kvPair) => kvPair.reverse())
-);
-
-const DEFAULT_VARIANT_TRACK_SERVER_ABBR = VariantTrackAbbr.RG;
+const DEFAULT_VARIANT_TRACK_SERVER_ABBR = TrackSourceAbbr.RG;
 const DEFAULT_VARIANT_TRACK_PVAL_COL = '7';
 const DEFAULT_VARIANT_TRACK_PPROB_COL = '8';
 
+export const DEFAULT_VARIANT_TRACK_SERVER = 'https://resgen.io/api/v1';
+export const DEFAULT_VARIANT_TRACK_TILESET = 'VF5-RDXWTxidGMJU7FeaxA';
 export const DEFAULT_VARIANT_TRACKS: VariantTrack[] = [
   {
-    server: 'https://resgen.io/api/v1',
-    tilesetUid: 'VF5-RDXWTxidGMJU7FeaxA',
+    server: DEFAULT_VARIANT_TRACK_SERVER,
+    tilesetUid: DEFAULT_VARIANT_TRACK_TILESET,
     columnPvalue: 7,
     columnPosteriorProbability: 8,
     markColor: 'black',
@@ -47,10 +38,12 @@ export const DEFAULT_VARIANT_TRACKS: VariantTrack[] = [
   },
 ];
 
-// TODO: Add strict type definition
-export const DEFAULT_VARIANT_TRACK_DEF = {
-  type: 'point-annotation',
+export const DEFAULT_VARIANT_TRACK_DEF: Track = {
+  type: TrackType.POINT_ANNOTATION,
+  uid: 'ibd-snps',
   height: 32,
+  server: DEFAULT_VARIANT_TRACK_SERVER,
+  tilesetUid: DEFAULT_VARIANT_TRACK_TILESET,
   options: {
     axisPositionHorizontal: 'right',
     markColor: 'black',
@@ -59,7 +52,8 @@ export const DEFAULT_VARIANT_TRACK_DEF = {
     markOpacity: 0.33,
     markOpacityFocus: 0.66,
     valueColumn: 7,
-    name: 'Variants',
+    focusRegion: [1680373143 + 81046453 - 25, 1680373143 + 81046453 + 25],
+    name: 'IBD Variants',
     labelPosition: 'topLeft',
     labelColor: 'black',
     labelOpacity: 0.33,
@@ -99,7 +93,7 @@ export function variantTracksEncoder(v: VariantTrack[]): string {
   }
 
   const track = v[0];
-  const serverAbbr = VARIANT_TRACK_SERVER_TO_ABBR[track.server];
+  const serverAbbr = SERVER_URL_TO_TRACK_SOURCE_ABBR[track.server];
 
   if (!track.tilesetUid || !serverAbbr) {
     throw new Error(
@@ -134,7 +128,7 @@ function variantTracksDecoder(v?: string): VariantTrack[] {
     columnPosteriorProbability = DEFAULT_VARIANT_TRACK_PPROB_COL,
   ] = v.split(':');
 
-  const server = VARIANT_TRACK_ABBR_TO_SERVER[serverAbbr as VariantTrackAbbr];
+  const server = TRACK_SOURCE_ABBR_TO_SERVER_URL[serverAbbr as TrackSourceAbbr];
 
   if (tilesetUid === undefined) return tilesetUid;
